add --list-highlight-themes command-line switch to show list of available highlight.js temes
